Start RAUC hawkBit client project
HawkBit is a provisioning server distributing software
bundles for multiple targets. The RAUC HawkBit client
communicates with the Direct Device Integration API.
If a deployment is available for the device the client
downloads the artifact, checks the MD5sum and triggers
RAUC install via DBUS. Status information are sent to
the provisioning server.

Signed-off-by: Bastian Stender <bst@pengutronix.de>
Signed-off-by: Enrico Jorns <ejo@pengutronix.de>