Added Option to use GatewayToken
Added a option if the target should authenticate through a GatewayToken
or a TargetToken, see
https://www.eclipse.org/hawkbit/concepts/authentication/ for more
information about the different "Token-Types".

The Type can be determined by adding the following option to the
config-File of the client:
auth_token_type = $TYPE
The folowing Types can be used: GatewayToken, TargetToken

If the option is not specified it will not break the program, instead it
tries to authenticate via TargetToken. This is to prevent errors when
updating the program with existing configs.

Signed-off-by: Max Hampel <hampel5275@gmail.com>